+++ lib/base.js
@@ -82,7 +82,6 @@ var Base = module.exports = function Base(args, options) {
 
   this.env = this.options.env;
   this.resolved = this.options.resolved;
-  this.fallbacks = this.options.generators || this.options.generator || {};
   this.generatorName = this.options.name || '';
 
   require('./env').enforceUpdate(this.env);
@@ -382,27 +381,19 @@ Base.prototype.run = function run(args, cb) {
  */
 
 Base.prototype.runHooks = function runHooks(cb) {
-  var self = this;
-  var hooks = this._hooks;
-
   cb = _.isFunction(cb) ? cb : function () {};
 
-  var callback = function (err) {
-    cb(err);
-  };
-
-  var resolve = function (hook) {
-    var resolved = self.defaultFor(hook.name);
-    var context = hook.as || self.resolved || self.generateName;
-    var options = _.clone(hook.options || self.options);
-    options.args = _.clone(hook.args || self.args);
+  var setupInvoke = function (hook) {
+    var resolved = this.defaultFor(hook.name);
+    var options = _.clone(hook.options || this.options);
+    options.args = _.clone(hook.args || this.args);
 
     return function (next) {
-      self.invoke(resolved + (context ? ':' + context : ''), options, next);
-    };
-  };
+      this.invoke(resolved + (hook.as ? ':' + hook.as : ''), options, next);
+    }.bind(this);
+  }.bind(this);
 
-  async.series(hooks.map(resolve), callback);
+  async.series(this._hooks.map(setupInvoke), cb);
 
   return this;
 };
@@ -419,9 +410,10 @@ Base.prototype.runHooks = function runHooks(cb) {
  *
  * ### Options:
  *
- *   - `as` The context value to use when runing the hooked generator
- *   - `args` The array of positional arguments to init and run the generator with
- *   - `options` An object containing a nested `options` property with the hash of options to use to init and run the generator with
+ *   - `as`      The context value to use when runing the hooked generator
+ *   - `args`    The array of positional arguments to init and run the generator with
+ *   - `options` An object containing a nested `options` property with the hash of options
+ *               to use to init and run the generator with
  *
  * ### Examples:
  *
@@ -477,15 +469,7 @@ Base.prototype.hookFor = function hookFor(name, config) {
  */
 
 Base.prototype.defaultFor = function defaultFor(name) {
-  var config = this.fallbacks;
-
-  if (this.options[name]) {
-    name = this.options[name];
-  } else if (config && config[name]) {
-    name = config[name];
-  }
-
-  return name;
+  return this.options[name] || name;
 };
 
 /**
